Författare ok

+++ MurvasBokhandel/Controllers/AuthorAdminController.cs
@@ -47,6 +47,8 @@ namespace MurvasBokhandel.Controllers
                 {
                     AuthorService.UpdateAuthor(a.Author);
 
+                    Auth.PushAlert(AlertView.Build("Författare uppdaterad.", AlertType.Success));
+
                     return View(AuthorService.GetAuthorWithBooksAndBooks(a.Author.Aid));
                 }
 
@@ -76,6 +78,8 @@ namespace MurvasBokhandel.Controllers
                 {
                     AuthorService.StoreAuthor(a);
 
+                    Auth.PushAlert(AlertView.Build("Författare "+a.FirstName + " "+ a.LastName +" är skapad.", AlertType.Success));
+
                     return RedirectToAction("Start");
                 }
 
@@ -90,6 +94,14 @@ namespace MurvasBokhandel.Controllers
         {
             if (Auth.HasAdminPermission())
             {
+                if (BookService.GetBooksByAuthor(a.Author.Aid).Count > 0)
+                {
+                    Auth.PushAlert(AlertView.Build("Kontrollera att ingen bok finns registrerad på författaren.", AlertType.Danger));
+
+                    return RedirectToAction("Author", new { id = a.Author.Aid });
+                }
+
+                Auth.PushAlert(AlertView.Build("Författare med Aid: " + a.Author.Aid + " är nu borttagen.", AlertType.Success));
                 AuthorService.DeleteAuthor(a.Author);
 
                 return RedirectToAction("Start");
